requirements.txt
agentic-doc>=0.2.4
python-dotenv>=1.0.0
opencv-python>=4.8.0
pydantic>=2.0.0 